requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI and LLM
openai==1.10.0
anthropic==0.18.0
langchain==0.1.6
langchain-openai==0.0.5

# Document Processing
PyPDF2==3.0.1
pypdf==4.0.1
python-docx==1.1.0
python-pptx==0.6.23
Pillow==10.2.0

# OCR
pytesseract==0.3.10
pdf2image==1.17.0

# Text Processing
beautifulsoup4==4.12.3
markdownify==0.11.6

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
aiofiles==23.2.1
requests==2.31.0

# Data Handling
pandas==2.2.0
numpy==1.26.3

# CORS
fastapi-cors==0.0.6
